requirements.txt
pyyaml 
requests 
minio
python-dotenv
faker 
psycopg2-binary
urllib3==2.0.7
confluent-kafka
black
flask